- OAuthFlowDirector.java: remove baseUri from the constructor parameters - Descendants of OAuthFlowDirector.java: Hardcode base urls, since they will not change that often and if they do the director should probably be reviewed anyway

+++ src/main/java/nl/juraji/biliomi/io/web/oauthflow/grants/code/model/OAuthAccessTokenResponse.java
@@ -30,7 +30,7 @@ public class OAuthAccessTokenResponse {
     this.accessToken = accessToken;
   }
 
-  public int getExpiresIn() {
+  public long getExpiresIn() {
     return expiresIn;
   }
